Added warehouse autons
Here's what I need to do still:
More consistent vision. I need to change the positions of the rectangles, which I can do, but I also need to get the comparison color and threshold value more consistent, which I can't do in this lighting.
Finish parking in the warehouse. Because of my limited field setup, I haven't gotten that done. I've gotten as far as I could, but I need to wait until the last second, then go full power into the warehouse. Or for the warehouse only auton, just park as soon as possible.