More consistent vision
I haven't tested this on the robot yet, since I'm on my computer, but this should make vision much more consistent. Instead of comparing to a color, it uses the hue value of the rectangle and sees which is the closest to yellow.

We are also going to try to have a second camera on the robot so we can see all positions in all scenarios.